+++ CHANGELOG.md
@@ -125,7 +125,7 @@
 
 # 0.5.1 (2020-10-30)
 
-- Ractor out lol-test as a crate.
+- Factor out lol-test as a crate.
 - Wait for noop entry to be committed before appending new entries in the new term.
 - Add AddServer/RemoveServer APIs to proto file. Deprecate InitCluster API.